move rotary files into shared/rotary
