+++ frontend/src/modules/shared/rotary/Knob.tsx
