+++ frontend/src/modules/shared/rotary/RotaryEncodeController.tsx
@@ -8,7 +8,7 @@ import {
   // getRotaryEncoderButtonPressed,
   getRotaryEncoderStep,
   getRotaryEncoderStepDiff,
-} from '../../store/controller/selectors';
+} from '../../../store/controller/selectors';
 
 /**
  * @typedef Props
